pyproject.toml
[project]
name = "streamlit-image-process-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "altair>=5.5.0",
    "attrs>=25.3.0",
    "blinker>=1.9.0",
    "cachetools>=5.5.2",
    "certifi>=2025.4.26",
    "charset-normalizer>=3.4.2",
    "click>=8.1.8",
    "colorama>=0.4.6",
    "contourpy>=1.3.2",
    "cycler>=0.12.1",
    "entrypoints>=0.4",
    "fonttools>=4.57.0",
    "gitdb>=4.0.12",
    "gitpython>=3.1.44",
    "greenlet>=3.2.1",
    "idna>=3.10",
    "jinja2>=3.1.6",
    "jsonschema>=4.23.0",
    "jsonschema-specifications>=2025.4.1",
    "kiwisolver>=1.4.8",
    "markdown-it-py>=3.0.0",
    "markupsafe>=3.0.2",
    "matplotlib>=3.10.1",
    "mdurl>=0.1.2",
    "narwhals>=1.37.1",
    "numpy>=2.2.5",
    "opencv-python-headless>=4.11.0.86",
    "packaging>=24.2",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "protobuf>=6.30.2",
    "pyarrow>=20.0.0",
    "pydeck>=0.9.1",
    "pygments>=2.19.1",
    "pyparsing>=3.2.3",
    "python-dateutil>=2.9.0.post0",
    "python-decouple>=3.8",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "referencing>=0.36.2",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "rpds-py>=0.24.0",
    "six>=1.17.0",
    "smmap>=5.0.2",
    "sqlalchemy>=2.0.40",
    "streamlit>=1.45.0",
    "streamlit-aggrid>=1.1.4.post1",
    "streamlit-option-menu>=0.4.0",
    "tenacity>=9.1.2",
    "toml>=0.10.2",
    "toolz>=1.0.0",
    "tornado>=6.4.2",
    "typing-extensions>=4.13.2",
    "tzdata>=2025.2",
    "urllib3>=2.4.0",
    "watchdog>=6.0.0",
]
